% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{as_int_group}
\alias{as_int_group}
\title{as_int_group, if nan/inf, then group return -1}
\usage{
as_int_group(group)
}
\arguments{
\item{group}{vector of double}
}
\description{
as_int_group, if nan/inf, then group return -1
}
