% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_qtl_turnover}
\alias{calc_qtl_turnover}
\title{calc_qtl_turnover
analyzing the turnover properties of a factor}
\usage{
calc_qtl_turnover(dt, qtl_name, shift)
}
\arguments{
\item{dt}{data table like struct}

\item{qtl_name}{default "f_qtl"}

\item{shift}{period default 1}
}
\description{
calc_qtl_turnover
analyzing the turnover properties of a factor
}
