% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_rank_acf}
\alias{calc_rank_acf}
\title{calc_rank_acf
Computes autocorrelation of factor ranks}
\usage{
calc_rank_acf(dt, f_name, shift)
}
\arguments{
\item{dt}{data table like struct}

\item{f_name}{default "factor"}

\item{shift}{period default 1}
}
\description{
calc_rank_acf
Computes autocorrelation of factor ranks
}
