% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_factor.R
\name{compute_mean_returns_spread}
\alias{compute_mean_returns_spread}
\title{compute_mean_returns_spread
compute the difference between the mean returns of quantiles, max - min}
\usage{
compute_mean_returns_spread(ret_dt, sd_dt)
}
\arguments{
\item{ret_dt}{dt, expect from FF::mean_rate_ret}

\item{sd_dt}{expect from FF::sd_rate_conversion}
}
\description{
compute_mean_returns_spread
compute the difference between the mean returns of quantiles, max - min
}
