% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_factor.R
\name{demean_forward_returns}
\alias{demean_forward_returns}
\title{demean_forward_returns}
\usage{
demean_forward_returns(dt, grouper = NA)
}
\arguments{
\item{dt}{table}

\item{grouper}{default NA, "f_qtl"}
}
\description{
demean_forward_returns
}
