% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_factor.R
\name{dummy_factor}
\alias{dummy_factor}
\title{dummy_factor}
\usage{
dummy_factor(n_ukey = 10L, n_day = 100L, n_qtls = 2L)
}
\arguments{
\item{n_ukey}{default 10L}

\item{n_day}{default 100L}

\item{n_qtls}{default 2L}
}
\description{
dummy_factor
}
