% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_factor.R
\name{factor_alpha_beta}
\alias{factor_alpha_beta}
\title{factor_alpha_beta
compute the alpha (excess returns), alpha t-stat (alpha significance), and beta (market exposure) of a factor.}
\usage{
factor_alpha_beta(
  factor_data,
  return_dt = NA,
  long_short = TRUE,
  equal_weight = FALSE
)
}
\arguments{
\item{factor_data}{dt, expect (group, factor, y_cols) format}

\item{return_dt}{expect FF::factor_returns format}

\item{long_short}{default TRUE}

\item{equal_weight}{default FALSE}
}
\description{
factor_alpha_beta
compute the alpha (excess returns), alpha t-stat (alpha significance), and beta (market exposure) of a factor.
}
