% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_IC.R
\name{factor_ics}
\alias{factor_ics}
\title{factor_ics
Computes the Spearman Rank Correlation based Information Coefficient (IC)}
\usage{
factor_ics(dt, group_adjust = FALSE, by_group = FALSE)
}
\arguments{
\item{group_adjust}{default FALSE}

\item{by_group}{default FALSE}
}
\description{
factor_ics
Computes the Spearman Rank Correlation based Information Coefficient (IC)
}
