% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_turnover.R
\name{factor_rank_acf}
\alias{factor_rank_acf}
\title{factor_rank_acf
Computes autocorrelation of factor ranks}
\usage{
factor_rank_acf(dt, f_name = "factor", shifts = c(1L))
}
\arguments{
\item{dt}{dt, expect factor data}

\item{f_name}{default "factor"}

\item{shifts}{period default 1}
}
\description{
factor_rank_acf
Computes autocorrelation of factor ranks
}
