% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_factor.R
\name{factor_returns}
\alias{factor_returns}
\title{factor_returns
calculate factor return by ticktime}
\usage{
factor_returns(dt, demean = TRUE, equal_weight = FALSE)
}
\arguments{
\item{demean}{default TRUE}

\item{equal_weight}{default FALSE}
}
\description{
factor_returns
calculate factor return by ticktime
}
