% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_turnover.R
\name{factor_turnovers}
\alias{factor_turnovers}
\title{factor_turnovers
analyzing the turnover properties of a factor}
\usage{
factor_turnovers(dt, qtl_name = "f_qtl", shifts = c(1L))
}
\arguments{
\item{dt}{dt, expect factor data}

\item{qtl_name}{default "f_qtl"}

\item{shifts}{period array default 1}
}
\description{
factor_turnovers
analyzing the turnover properties of a factor
}
