% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-adventures-bestiary-data.R
\docType{data}
\name{Final_Fantasy_Adventure}
\alias{Final_Fantasy_Adventure}
\title{Final Fantasy Adventure Bestiary data}
\format{
\code{Final_Fantasy_Adventure}
A data frame with 98 rows and 6 columns
\describe{
\item{Name}{Name}
\item{HP}{Hit Points}
\item{AP}{Attack Power}
\item{Defense Power}{Defense Power}
\item{Exp}{Experience}
\item{Gil}{Gil}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Final_Fantasy_Adventure}
}
\usage{
Final_Fantasy_Adventure
}
\description{
Data on all monsters in Final Fantasy Adventure
}
\keyword{datasets}
