% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-I-bestiary-data.R
\docType{data}
\name{final_fantasy_I_bestiary_data}
\alias{final_fantasy_I_bestiary_data}
\title{Final Fantasy One Bestiary data}
\format{
\code{final_fantasy_I_bestiary_data}
A data frame with 128 rows and 11 columns
\describe{
\item{Name}{Name}
\item{HP}{Hit Point}
\item{Defense}{Defense}
\item{Evasion}{Evasion}
\item{Magic Defense}{Magic Defense}
\item{Attack}{Attack}
\item{Accuracy}{Accuracy}
\item{Hit#}{Hit Number}
\item{Crit}{Critical}
\item{EXP}{Experience Point}
\item{Gil}{Gil}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Bestiary_(Final_Fantasy)}
}
\usage{
final_fantasy_I_bestiary_data
}
\description{
Data on all monsters in Final Fantasy 1
}
\keyword{datasets}
