% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-II-bestiary-data.R
\docType{data}
\name{Final_Fantasy_II}
\alias{Final_Fantasy_II}
\title{Final Fantasy Two Bestiary data}
\format{
\code{Final_Fantasy_II}
A data frame with 128 rows and 10 columns
\describe{
\item{Name}{Name}
\item{Rank}{Rank}
\item{HP}{Hit Points}
\item{MP}{Magic Points}
\item{Attack}{Attacks}
\item{Defense}{Defense}
\item{Magic Defense}{Magic Defense}
\item{Accuracy}{Accuracy}
\item{Evasion}{Evasion}
\item{Gil}{Gil}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Bestiary_(Final_Fantasy_II)}
}
\usage{
Final_Fantasy_II
}
\description{
Data on all monsters in Final Fantasy 2
}
\keyword{datasets}
