% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-III-bestiary-data.R
\docType{data}
\name{Final_Fantasy_III}
\alias{Final_Fantasy_III}
\title{Final Fantasy Three Bestiary data}
\format{
\code{Final_Fantasy_III}
A data frame with 226 rows and 7 columns
\describe{
\item{Name}{Name}
\item{Level}{Level}
\item{HP}{Hit Points}
\item{Attack}{Attack}
\item{Defense}{Defense}
\item{EXP}{Experience Points}
\item{Gil}{Gil}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Bestiary_(Final_Fantasy_III)}
}
\usage{
Final_Fantasy_III
}
\description{
Data on all monsters in Final Fantasy 3
}
\keyword{datasets}
