% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-IV-bestiary-data.R
\docType{data}
\name{Final_Fantasy_IV}
\alias{Final_Fantasy_IV}
\title{Final Fantasy Four Bestiary data}
\format{
\code{Final_Fantasy_IV}
A data frame with 197 rows and 15 columns
\describe{
\item{Name}{Name}
\item{Form}{Form}
\item{Level}{Level}
\item{HP}{Hit Points}
\item{Attack}{Attack}
\item{Defense}{Defense}
\item{Magic}{Magic}
\item{Attack\%}{Attack Percentage}
\item{Defense\%}{Defense Percentage}
\item{Magic Def\%}{Magic Defense Percentage}
\item{Magic Def}{Magic Defense}
\item{Agility}{Agility}
\item{EXP}{Experience Points}
\item{Gil}{Gil}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Bestiary_(Final_Fantasy_IV)#Final_Fantasy_IV_Pixel_Remaster}
}
\usage{
Final_Fantasy_IV
}
\description{
Data on all monsters in Final Fantasy 4
}
\keyword{datasets}
