% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-IX-bestiary-data.R
\docType{data}
\name{Final_Fantasy_IX}
\alias{Final_Fantasy_IX}
\title{Final Fantasy Nine Bestiary data}
\format{
\code{Final_Fantasy_IX}
A data frame with 197 rows and 17 columns
\describe{
\item{Name}{Name}
\item{Form}{Form}
\item{LV}{Level}
\item{HP}{Hit Points}
\item{MP}{Magic Points}
\item{Strength}{Strength}
\item{Defense}{Defense}
\item{Evade}{Evade}
\item{Magic}{Magic}
\item{Magic Def}{Magic Defense}
\item{Magic Eva}{Magic Evasion}
\item{Attack}{Attack}
\item{Spirit}{Spirit}
\item{Speed}{Speed}
\item{Exp}{Experience Point}
\item{Gil}{Gil}
\item{AP}{Ability Point}

}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Monster_(Final_Fantasy_IX)}
}
\usage{
Final_Fantasy_IX
}
\description{
Data on all monsters in Final Fantasy 9
}
\keyword{datasets}
