% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_fantasy_one_bestiary_data.R
\docType{data}
\name{final_fantasy_one_bestiary_data}
\alias{final_fantasy_one_bestiary_data}
\title{Final Fantasy One Bestiary data}
\format{
\code{final_fantasy_one_bestiary_data}
A data frame with 128 rows and 11 columns
\describe{
\item{name}{Name of the beast}
\item{HP}{How many hit points they have}
\item{Defense}{How do they best against physical attacks}
\item{Evasion}{How fast can they dodge your attacks}
\item{Magic Defense}{How do they best against magical attacks}
\item{Attack}{How strong are they're attacks}
\item{Accuracy}{How precise they hit you}
\item{Hit#}{what's the max hit they can do}
\item{Crit}{What's the percentage it'll crit you}
\item{EXP}{How much exp you earn}
\item{Gil}{How much money the defeated enemy drops}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Bestiary_(Final_Fantasy)}
}
\usage{
final_fantasy_one_bestiary_data
}
\description{
Data on all monsters in Final Fantasy 1
}
\keyword{datasets}
