% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-VI-bestiary-data.R
\docType{data}
\name{Final_Fantasy_VI}
\alias{Final_Fantasy_VI}
\title{Final Fantasy Six Bestiary data}
\format{
\code{Final_Fantasy_VI}
A data frame with 347 rows and 15 columns
\describe{
\item{Name}{Name}
\item{Form}{Form}
\item{LV}{Level}
\item{HP}{Hit Points}
\item{MP}{Magic Points}
\item{Attack}{Attack}
\item{Magic}{Magic}
\item{Defense}{Defense}
\item{Magic Defense}{Magic Defense}
\item{Magic Evasion}{Magic Evasion}
\item{Agility}{Agility}
\item{Hit Rate}{Hit Rate}
\item{Evasion}{Evasion}
\item{EXP}{Experience Points}
\item{Gil}{Gil}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Bestiary_(Final_Fantasy_VI)#Bestiary_(pixel_remaster)}
}
\usage{
Final_Fantasy_VI
}
\description{
Data on all monsters in Final Fantasy 6
}
\keyword{datasets}
