% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-VII-bestiary-data.R
\docType{data}
\name{Final_Fantasy_VII}
\alias{Final_Fantasy_VII}
\title{Final Fantasy Seven Bestiary data}
\format{
\code{Final_Fantasy_VII}
A data frame with 290 rows and 15 columns
\describe{
\item{Name}{Name}
\item{Form}{Form}
\item{LV}{Level}
\item{HP}{Hit Points}
\item{MP}{Magic Points}
\item{Attack}{Attack}
\item{Magic Attack}{Magic Attack}
\item{Defense}{Defense}
\item{Magic Defense}{Magic Defense}
\item{Dexterity}{Dexterity}
\item{Defense\%}{Defense Percentage}
\item{Luck}{Luck}
\item{EXP}{Experinece Points}
\item{AP}{Ability Points}
\item{Gil}{Gil}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Final_Fantasy_VII_enemies}
}
\usage{
Final_Fantasy_VII
}
\description{
Data on all monsters in Final Fantasy 7
}
\keyword{datasets}
