% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-VIII-bestiary-data.R
\docType{data}
\name{Final_Fantasy_VIII}
\alias{Final_Fantasy_VIII}
\title{Final Fantasy Eight Bestiary data}
\format{
\code{Final_Fantasy_VIII}
A data frame with 128 rows and 11 columns
\describe{
\item{Name}{Name}
\item{Form}{Form}
\item{LV}{Level}
\item{HP}{Hit Points}
\item{STR}{Strength}
\item{MAG}{Magic}
\item{VIT}{Vitality}
\item{SPR}{Spirit}
\item{SPD}{Speed}
\item{EVA}{Evasion}
\item{EXP}{Experience Points}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Final_Fantasy_VIII_enemies}
}
\usage{
Final_Fantasy_VIII
}
\description{
Data on all monsters in Final Fantasy 8
}
\keyword{datasets}
