% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-VIII-bestiary-data.R
\docType{data}
\name{final_fantasy_VIII_bestiary_data}
\alias{final_fantasy_VIII_bestiary_data}
\title{Final Fantasy Eight Bestiary data}
\format{
\code{final_fantasy_VIII_bestiary_data}
A data frame with 128 rows and 10 columns
\describe{
\item{Name}{Name}
\item{LV}{Level}
\item{HP}{Hit Points}
\item{STR}{Strength}
\item{MAG}{Magic}
\item{VIT}{Vitality}
\item{SPR}{Spirit}
\item{SPD}{Speed}
\item{EVA}{Evasion}
\item{EXP}{Experience Points}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Final_Fantasy_VIII_enemies}
}
\usage{
final_fantasy_VIII_bestiary_data
}
\description{
Data on all monsters in Final Fantasy 8
}
\keyword{datasets}
