% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-X-bestiary-data.R
\docType{data}
\name{Final_Fantasy_X}
\alias{Final_Fantasy_X}
\title{Final Fantasy Ten Bestiary data}
\format{
\code{Final_Fantasy_X}
A data frame with 246 rows and 14 columns
\describe{
\item{Name}{Name}
\item{Form}{Form}
\item{HP}{Hit Points}
\item{MP}{Magic Points}
\item{AP}{Ability Points}
\item{Gil}{Gil}
\item{Strength}{Strength}
\item{Magic}{Magic}
\item{Defense}{Defense}
\item{Magic Defense}{Magic Defense}
\item{Agility}{Agility}
\item{Accuracy}{Accuracy}
\item{Evasion}{Evasion}
\item{Luck}{Luck}
}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Sphere_of_Monster}
}
\usage{
Final_Fantasy_X
}
\description{
Data on all monsters in Final Fantasy 10
}
\keyword{datasets}
