% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-XII-bestiary-data.R
\docType{data}
\name{final_fantasy_XII_bestiary_data}
\alias{final_fantasy_XII_bestiary_data}
\title{Final Fantasy Twelve Bestiary data}
\format{
\code{final_fantasy_XII_bestiary_data}
A data frame with 620 rows and 17 columns
\describe{
\item{Name}{Name}
\item{Form}{Form}
\item{LV}{Level}
\item{HP}{Health Points}
\item{MP}{Magic Points}
\item{Strength}{Strength}
\item{Magick Power}{Magick Power}
\item{Vitality}{Vitality}
\item{Speed}{Speed}
\item{Attack Power}{Attack Power}
\item{Defense}{Defense}
\item{Magick Resist}{Magick Resist}
\item{Evade}{Evade}
\item{EXP}{Experience Points}
\item{LP}{License Points}
\item{CP}{Clan Points}
\item{Gil}{Gil}

}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Bestiary_(Final_Fantasy_XII)}
}
\usage{
final_fantasy_XII_bestiary_data
}
\description{
Data on all monsters in Final Fantasy 12
}
\keyword{datasets}
