% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-XIII-bestiary-data.R
\docType{data}
\name{Final_Fantasy_XIII}
\alias{Final_Fantasy_XIII}
\title{Final Fantasy Thirteen Bestiary data}
\format{
\code{Final_Fantasy_XIII}
A data frame with 244 rows and 12 columns
\describe{
\item{Name}{Name}
\item{Form}{Form}
\item{Level}{Level}
\item{HP}{Health Points}
\item{Strength}{Strength}
\item{Magic}{Magic}
\item{Stagger}{Stagger}
\item{Chain Resistance}{Chain Resistance}
\item{CP}{Clan Points}
\item{Keep}{Keep}
\item{Launch}{Launch}
\item{Libra}{Libra}

}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Datalog_(Final_Fantasy_XIII)/Enemy_Intel}
}
\usage{
Final_Fantasy_XIII
}
\description{
Data on all monsters in Final Fantasy 13
}
\keyword{datasets}
