% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final-fantasy-XV-bestiary-data.R
\docType{data}
\name{Final_Fantasy_XV}
\alias{Final_Fantasy_XV}
\title{Final Fantasy Fifthteen Bestiary data}
\format{
\code{Final_Fantasy_XV}
A data frame with 362 rows and 11 columns
\describe{
\item{Name}{Name}
\item{Form}{Form}
\item{Level}{Level}
\item{HP}{Health Points}
\item{Size}{Size}
\item{EXP}{Experience Points}
\item{Libra}{Libra}
\item{Strength}{Strength}
\item{Vitality}{Vitality}
\item{Parry}{Parry}
\item{Spirit}{Spirit}

}
}
\source{
\url{https://finalfantasy.fandom.com/wiki/Bestiary_(Final_Fantasy_XV)}
}
\usage{
Final_Fantasy_XV
}
\description{
Data on all monsters in Final Fantasy 15
}
\keyword{datasets}
