% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_factor.R
\name{get_forward_returns_columns}
\alias{get_forward_returns_columns}
\title{get_forward_returns_columns}
\usage{
get_forward_returns_columns(dt)
}
\arguments{
\item{dt}{table}
}
\description{
get_forward_returns_columns
}
