% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_factor.R
\name{get_return_table}
\alias{get_return_table}
\title{get_return_table}
\usage{
get_return_table(alpha_beta, mean_qtl_rate_ret, return_diff)
}
\arguments{
\item{alpha_beta}{dt, expect from FF::factor_alpha_beta}

\item{mean_qtl_rate_ret}{expect from FF::mean_rate_ret}

\item{return_diff}{expect from FF::compute_mean_returns_spread}
}
\description{
get_return_table
}
