\name{Internal_ffdf}
\Rdversion{1.1}
\alias{get_nvw}
\title{
  Internal ffdf functions
}
\description{
  Internal ffdf functions
}
\usage{
get_nvw(x)
}
\arguments{
  \item{x}{an object}
}
\details{
'get_nvw' gives the length and vw of an object (or its first dimension, if it is not a vector)
}
\keyword{ internal }
