\name{is.ff}
\alias{is.ff}
\title{ Test for class ff }
\description{
  checks if x inherits from class "ff"
}
\usage{
is.ff(x)
}
\arguments{
  \item{x}{ any object }
}
\value{
  logical scalar
}
\author{ Jens Oehlschlägel }
\seealso{  \code{\link{inherits}}, \code{\link{as.ff}}, \code{\link{is.ffdf}} }
\examples{
  is.ff(integer())
}
\keyword{ IO }
\keyword{ data }
