\name{is.ffdf}
\alias{is.ffdf}
\title{ Test for class ff }
\description{
  checks if x inherits from class "ffdf"
}
\usage{
is.ffdf(x)
}
\arguments{
  \item{x}{ any object }
}
\value{
  logical scalar
}
\author{ Jens Oehlschlägel }
\seealso{  \code{\link{inherits}}, \code{\link{as.ffdf}}, \code{\link{is.ff}} }
\examples{
  is.ffdf(integer())
}
\keyword{ IO }
\keyword{ data }
