% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_factor.R
\name{mean_return_by_quantile}
\alias{mean_return_by_quantile}
\title{mean_return_by_quantile}
\usage{
mean_return_by_quantile(dt, demeaned = TRUE, by_date = FALSE)
}
\arguments{
\item{dt}{table}

\item{demeaned}{default TRUE}

\item{by_date}{default FALSE}
}
\description{
mean_return_by_quantile
}
