% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_chart.R
\name{plot_cum_rets}
\alias{plot_cum_rets}
\title{plot_cum_rets}
\usage{
plot_cum_rets(dt, first_n = 0L, last_n = 0L)
}
\arguments{
\item{dt}{expect from FF::cumret_dt}

\item{first_n}{0L}

\item{last_n}{0L}
}
\description{
plot_cum_rets
}
