% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_chart.R
\name{plot_cum_rets_by_qtl}
\alias{plot_cum_rets_by_qtl}
\title{plot_cum_rets_by_qtl}
\usage{
plot_cum_rets_by_qtl(dt, period = "y1d", first_n = 0L, last_n = 0L)
}
\arguments{
\item{dt}{expect from FF::mean_rate_ret}

\item{period}{default y1d}

\item{first_n}{0L}

\item{last_n}{0L}
}
\description{
plot_cum_rets_by_qtl
}
