% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_chart.R
\name{plot_ic_hist}
\alias{plot_ic_hist}
\title{plot_ic_hist}
\usage{
plot_ic_hist(dt, n_bins = 10L, use_density_hight = TRUE)
}
\arguments{
\item{dt}{expect from FF::factor_ics}

\item{n_bins}{default to 10 bins}

\item{use_density_hight}{show density}
}
\description{
plot_ic_hist
}
