% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_chart.R
\name{plot_ic_ts}
\alias{plot_ic_ts}
\title{plot_ic_ts}
\usage{
plot_ic_ts(dt, MA_LEN = 3L, first_n = 0L, last_n = 0L)
}
\arguments{
\item{dt}{expect from FF::factor_ics}

\item{MA_LEN}{3L}

\item{first_n}{0L}

\item{last_n}{0L}
}
\description{
plot_ic_ts
}
