% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_chart.R
\name{plot_quantile_returns_bar}
\alias{plot_quantile_returns_bar}
\title{plot_quantile_returns_bar}
\usage{
plot_quantile_returns_bar(dt)
}
\arguments{
\item{dt}{expect from FF::mean_rate_ret}
}
\description{
plot_quantile_returns_bar
}
