% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_chart.R
\name{plot_quantile_returns_violin}
\alias{plot_quantile_returns_violin}
\title{plot_quantile_returns_violin}
\usage{
plot_quantile_returns_violin(dt)
}
\arguments{
\item{dt}{expect from FF::mean_rate_ret by date}
}
\description{
plot_quantile_returns_violin
}
