% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qcut}
\alias{qcut}
\title{quantile_cut, if nan/inf, factor value will be mapped to 0}
\usage{
qcut(data, numQtls)
}
\arguments{
\item{data}{vector of double}

\item{numQtls}{number of quantiles}
}
\description{
quantile_cut, if nan/inf, factor value will be mapped to 0
}
