% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_factor.R
\name{return_report}
\alias{return_report}
\title{return_report}
\usage{
return_report(dt, long_short = TRUE, group_neutral = FALSE, by_group = FALSE)
}
\arguments{
\item{dt}{dt, expect like FF::dummy_factor}
}
\description{
return_report
}
