% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FF_factor.R
\name{sd_rate_conversion}
\alias{sd_rate_conversion}
\title{sd_rate_conversion
adjust sd to base period}
\usage{
sd_rate_conversion(dt)
}
\arguments{
\item{dt}{table, expect (group, y_cols) format}
}
\description{
sd_rate_conversion
adjust sd to base period
}
