% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{time_group}
\alias{time_group}
\title{as_int_group, if nan/inf, then group return -1}
\usage{
time_group(date, name)
}
\arguments{
\item{name}{"d" for day, "m" for month, "y" for year}

\item{group}{vector of int}
}
\description{
as_int_group, if nan/inf, then group return -1
}
