% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{to_weights}
\alias{to_weights}
\title{to_weights}
\usage{
to_weights(v, demean, equal_weight)
}
\arguments{
\item{v}{vector of double}

\item{demean}{if vector demean}

\item{equal_weight}{if vector equal_weight}
}
\description{
to_weights
}
